FSM fixes, new logic preprocessing via states todo, criteria evaluation change

+++ src/main/java/cz/cvut/fel/mongodb_assignment_evaluator/domain/enums/ResultStates.java
@@ -12,7 +12,7 @@ public enum ResultStates {
 
     private final String text;
 
-    public static ResultStates evaluate(int currentScore, int requiredScore) {
+    public static ResultStates evaluate(long currentScore, long requiredScore) {
         if (currentScore <= 0) {
             return UNFULFILLED;
         } else if (currentScore >= requiredScore) {
